Change "base name" format to fix workflow on PR.
This changes the way the base name (used for naming artifacts and
release files) is formatted. The helper script will check if the
latest git tag begins with the version number from `src/params.h`,
which should be true under a normal workflow. If it is true, then
`git describe --tags` will be used for the base name. It should
include the version, short hash, and the number of commits since
the last tag when `git HEAD` isn't directly referenced by a tag.
Otherwise, if the current commit has a tag reference (i.e., it's a
release run), we can use shorter file names without the commit
hash for releases.